The only true initial commit
Commiting in parts was a bit hard due to stories not being isolated.
Will do this better next time. Oh, and all tests pass.

NOTE that a LOT of this code will be refactored. This shit is complex,
so let just let it emerge, a'ight?

- Added basic gem structure
- Added dummy implementation (called dummy)
- Added yardoc for proper documentation
- Added configuration object to manage dynamic rest client, memcached
  etc
- Implemented the find method as it will be supposed to work
- Added some basic cache handling
- Added helpers for determining file paths
- Added some simple request parsing logic
- Added fakeweb to stub out the rest client
- Added helper methods to make the rest client responses dynamic
- Redefined `==` on dada::base to not check for object id, but instance
  variables only (like AM)

There's a lot left to be done, but this proofs its damn well possible
:-)

Banana!