Refactored attributes
Refactored attributes to dynamic setters and accessors while all
attributes are maintained in a central hash. Rails was giving variable
trouble with attributes not being always set, so now we just store these
in one place.

Also, all attribute names are now also stored in _attributes. This might
be double, but this seems to be safer to me.

Note that right now the hash key/values only get set if a VALUE GETS
SET. For validation, this should NOT matter (nil is nil).

Added tests to check if attributes gets added to json properly.

If this shit keeps up I might test this against a real rails server.